+++ packages/flame/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Next]
  - Add `ButtonComponent` backed by two `PositionComponent`s
  - Add `SpriteButtonComponent` backed by two `Sprite`s
+ - Fix `Parallax.load` with different loading times
 
 ## [1.0.0-releasecandidate.18]
  - Forcing portrait and landscape mode is now supported on web